+++ ci/buildspec.yml
@@ -6,7 +6,7 @@ phases:
   install:
     commands:
       - apt-get update
-      - apt-get install -y curl
+      - apt-get install -y curl gnupg2
       - pip install pip -U
       - pip install future
       - pip install Pillow
@@ -20,7 +20,7 @@ phases:
   build:
     commands:
       - ./torchserve_sanity.sh
-      - cd serving-sdk/ && mvn clean deploy && cd ../
+      - cd serving-sdk/ && mvn clean install && cd ../
 
 artifacts:
   files:
