+++ frontend/gradle.properties
@@ -6,4 +6,4 @@ slf4j_log4j12_version=1.7.25
 gson_version=2.8.5
 commons_cli_version=1.3.1
 testng_version=6.8.1
-mms_server_sdk_version=1.0.1
+torchserve_sdk_version=0.0.3
