+++ frontend/server/build.gradle
@@ -2,7 +2,7 @@ dependencies {
     compile "io.netty:netty-all:${netty_version}"
     compile project(":modelarchive")
     compile "commons-cli:commons-cli:${commons_cli_version}"
-    compile "software.amazon.ai:mms-plugins-sdk:${mms_server_sdk_version}"
+    compile "org.pytorch:torchserve-plugins-sdk:${torchserve_sdk_version}"
     testCompile "org.testng:testng:${testng_version}"
 }
