+++ frontend/server/src/main/java/org/pytorch/serve/servingsdk/impl/ModelServerResponse.java
@@ -5,7 +5,7 @@ import io.netty.handler.codec.http.FullHttpResponse;
 import io.netty.handler.codec.http.HttpHeaderNames;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import java.io.OutputStream;
-import software.amazon.ai.mms.servingsdk.http.Response;
+import org.pytorch.serve.servingsdk.http.Response;
 
 public class ModelServerResponse implements Response {
