+++ frontend/server/src/main/java/org/pytorch/serve/servingsdk/impl/ModelWorker.java
@@ -1,8 +1,8 @@
 package org.pytorch.serve.servingsdk.impl;
 
+import org.pytorch.serve.servingsdk.Worker;
 import org.pytorch.serve.wlm.WorkerState;
 import org.pytorch.serve.wlm.WorkerThread;
-import software.amazon.ai.mms.servingsdk.Worker;
 
 public class ModelWorker implements Worker {
     private boolean running;
