+++ plugins/endpoints/src/main/resources/META-INF/services/org.pytorch.serve.servingsdk.ModelServerEndpoint
@@ -1,0 +1,1 @@
+org.pytorch.serve.plugins.endpoint.ExecutionParameters
