+++ plugins/gradle.properties
@@ -1,0 +1,2 @@
+gson_version=2.8.5
+torchserve_sdk_version=0.0.3
