+++ plugins/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Tue Jun 04 12:29:27 PDT 2019
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-4.10-all.zip
