+++ serving-sdk/src/main/java/org/pytorch/serve/servingsdk/annotations/helpers/EndpointTypes.java
@@ -1,0 +1,9 @@
+
+package org.pytorch.serve.servingsdk.annotations.helpers;
+
+/**
+ * Types of ModelServer endpoints
+ */
+public enum EndpointTypes {
+    NONE, INFERENCE, MANAGEMENT;
+}
