--- serving-sdk/src/main/java/software/amazon/ai/mms/servingsdk/ModelServerEndpointException.java
@@ -1,11 +1,0 @@
-
-
-package software.amazon.ai.mms.servingsdk;
-
-/**
- * Runtime exception for custom model server endpoint plugins
- */
-public class ModelServerEndpointException extends RuntimeException {
-    public ModelServerEndpointException(String err) {super(err);}
-    public ModelServerEndpointException(String err, Throwable t) {super(err, t);}
-}
