--- serving-sdk/src/main/java/software/amazon/ai/mms/servingsdk/annotations/helpers/EndpointTypes.java
@@ -1,9 +1,0 @@
-
-package software.amazon.ai.mms.servingsdk.annotations.helpers;
-
-/**
- * Types of ModelServer endpoints
- */
-public enum EndpointTypes {
-    NONE, INFERENCE, MANAGEMENT;
-}
